Try block, logging, edge cases, for loop, list
Added logging because there are too many edge cases to predict everything. 
Added a second database to import the list of games, generated using main-program.py(this will change slightly in the future to just use one).
Added a try block to the whole script because of edge cases breaking it.
Made number + name be primary key because some teams have 3 goalies, two of which share a number. 
Added some edge case detection to fix goalie number and minutes played.